// src/types/atp.ts
export interface AtpData {
    dinaskary: number;
    dinasmitra: number;
    dinasopr: number;
    ebca: number;
    ebni: number;
    ebri: number;
    edki: number;
    eflo: number;
    emandiri: number;
    emega: number;
    enobu: number;
    golongan: number;
    id: number;
    idasalgerbang: number;
    idcabang: number;
    idgardu: number;
    idgerbang: number;
    shift: number;
    tanggal: string;
    tunai: number;
  }